+++ benchmarks/messaging_patterns_benchmark.cpp
@@ -11,8 +11,8 @@
  */
 
 #include "../include/psyne/core/behaviors.hpp"
-#include "../include/psyne/global/logger.hpp"
-#include "../include/psyne/global/threadpool.hpp"
+#include <psyne/global/logger.hpp>
+#include <psyne/global/threadpool.hpp>
 #include <atomic>
 #include <chrono>
 #include <cstring>